refactor: rename packages for ms tests
